bump at-least version in version check
The t/librdkafka.t test file tests that the version is at least
the same as $VERSION for this dist.  #1 bumped the verion for a
share install, but didn't bump the at-least check for a system
install.  This makes the at-least chek the same as what is
required for the test to pass (though it may require a share
install on some systems with a slightly older librdkafka).

Alternatively, it may make sense to loosen the requirement.
The latest version of ubuntu 0.11.5 and could likely be
sufficient for most purposes, and would get OS secutirty
patches as necessary.