Download the latest (as of the writing) version
as of 12 days ago this is v1.1.0

Also remove --enable-static from the configure line.  Normally
this option would build a static library, however, configure
does not seem to be based on autoconf, and its behavior is
actually linking AGAINST static libraries instead.  On some
platforms this can cause a link fail if the system libraries
aren't built with -fPIC / -fpic

 /usr/bin/ld: /usr/lib/gcc/x86_64-linux-gnu/6/../../../x86_64-linux-gnu/libz.a(deflate.o): relocation R_X86_64_PC32 against symbol `z_errmsg' can not be used when making a shared object; recompile with -fPIC

This was not the case in the v0.11.6.  Without --enable-static
it still builds a .a file which can be used by XS and a .so
file which can be used by FFI.